feat: EVEN 모드 로직 단순화 (#266)
finalCount 계산 방식을 간소화하여 코드 복잡도 감소.
기존에 사용되던 determineFinalCount 메서드 제거.

+++ src/main/java/org/devkor/apu/saerok_server/domain/collection/core/repository/CollectionRepository.java
@@ -155,7 +155,7 @@ public class CollectionRepository {
             double radiusMeters,
             Long userId,
             boolean isMineOnly,
-            int limit,
+            long limit,
             double gridSizeMeters
     ) {
         if (limit <= 0) {
